Updates to Lincoln to make functions clearer
