+++ lincoln/layers.py
@@ -61,7 +61,8 @@ class Layer(object):
 
 class Dense(Layer):
     '''
-    Once we define all the Operations and the basi.
+    Once we define all the Operations and the outline of a layer, all that remains to implement here 
+    is the _setup_layer function!
     '''
     def __init__(self, 
                  neurons: int, 
@@ -80,82 +81,3 @@ class Dense(Layer):
                            BiasAdd(self.params[1])] + [self.activation]
 
         
-class LayerBlock(object):
-    '''
-    We will ultimately want another level on top of operations and Layers, for example when we get to ResNets.
-    For now, I'm calling that a "LayerBlock" and defining a "NeuralNetwork" to be identical to it. 
-    '''
-    def __init__(self, layers: List[Layer]):
-        super().__init__()
-        self.layers = layers
-              
-    def forward(self, X_batch: Tensor) -> Tensor:
-        
-        X_out = X_batch
-        for layer in self.layers:
-            X_out = layer.forward(X_out)
-
-        return X_out
-        
-    def backward(self, loss_grad: Tensor) -> Tensor:
-    
-        grad = loss_grad
-        for layer in reversed(self.layers):
-            grad = layer.backward(grad)
-            
-        return grad
-    
-    def params(self):
-        for layer in self.layers:
-            yield from layer.params()
-        
-    def param_grads(self):
-        for layer in self.layers:
-            yield from layer.param_grads()
-    
-    def __iter__(self):
-        return iter(self.layers)
-    
-    def __repr__(self):
-        layer_strs = [str(layer) for layer in self.layers]
-        return f"{self.__class__.__name__}(\n  " + ",\n  ".join(layer_strs) + ")"
-    
-
-from .losses import Loss, MeanSquaredError
-class NeuralNetwork(LayerBlock):
-    '''
-    Just a list of layers that runs forwards and backwards
-    '''
-    def __init__(self, layers: List[Layer], 
-                 learning_rate: float = 0.01, 
-                 loss: Loss = MeanSquaredError):
-        super().__init__(layers)
-        self.learning_rate = learning_rate
-        self.loss = loss
-        
-    def forward_loss(self, 
-                     X_batch: Tensor, 
-                     y_batch: Tensor) -> float:
-        
-        prediction = self.forward(X_batch)
-        return self.loss.loss_gradient(prediction, y_batch)
-        
-    def train_batch(self, 
-                    X_batch: Tensor,
-                    y_batch: Tensor) -> float:
-        
-        loss = self.forward_loss(X_batch, y_batch)
-        
-        self.backward(self.loss.loss_grad)
-        
-        self.update_params()
-         
-        return loss
-
-    def update_params(self) -> None:
-
-        for layer in self.layers:
-            
-            for param, grad in zip(layer.params, layer.param_grads):
-                param.sub_(self.learning_rate * grad)
-
